document-summarizer/requirements.txt
google-genai>=0.3.0
python-dotenv>=1.0.0
rich>=13.7.0
